Added routing to all pages
