+++ .gitignore
@@ -129,3 +129,7 @@ dmypy.json
 .pyre/
 
 secrets.toml
+
+
+# Jetbrains IDE
+.idea
